+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Bug / Security fixes:
 * Fix #207: Broken disc info on OGG files
 * Fix #199: Improve the message on the cli tool
 * Fix #196: Error on import when people mess with dates
+* Fix #192: Skip symlinks on the import process to avoid infinite loops
 * Fix #183: Bug if the artist string contains "$"
 * Fix #178: Problem with file rights at installation
 * Fix #177: Trim whitespace characters on search request